produce the same output but using objects
Parse the XML into Ruby objects and then unparse it to produce the same
output. Now that everything is in Ruby, it will be possible to generate
different output formats like JSON or the XML and JSON-LD currently
being served by the morphsvc Python code.

The data model is based on the structure of the XML from the
Perseids/Alpheios version of Morpheus. I'm not sure it's the right
choice; it feels very verbose. Another option would be to simply use
arrays and hashes but that feels too loosey-goosey. We can see how the
code evolves and revisit the data model after the JSON endpoint is
designed.