move redis caching to morpheus result
Move the Redis caching from the endpoints themselves to the calls to
Morpheus. This way the caching will still work whether the output is
given in JSON, XML, or something else.

There should still be caching done at the higher level, but there's no
reason to do this in the application itself. It's more efficient to put
it behind a proxy (nginx, Apache) and use that proxy's caching. At some
point I'll add nginx config to do that to the Docker Compose
configuration.

This commit also makes some changes to the way options are handled in
anticipation for multiple endpoints that will handle them differently.